+++ runtime-core/core-types/definition/mixed.cpp
@@ -13,8 +13,7 @@ void mixed::destroy() noexcept {
       as_array().~array<mixed>();
       break;
     case type::OBJECT: {
-      auto ptr = *reinterpret_cast<vk::intrusive_ptr<may_be_mixed_base>*>(&storage_);
-      ptr.~intrusive_ptr<may_be_mixed_base>();
+      reinterpret_cast<vk::intrusive_ptr<may_be_mixed_base>*>(&storage_)->~intrusive_ptr<may_be_mixed_base>();
       storage_ = 0;
       break;
     }
