+++ runtime-core/runtime-core-context.h
@@ -5,33 +5,11 @@
 #pragma once
 
 #include <cstddef>
-#include "runtime-core/memory-resource/unsynchronized_pool_resource.h"
 
 #ifndef INCLUDED_FROM_KPHP_CORE
   #error "this file must be included only from runtime-core.h"
 #endif
 
-struct RuntimeAllocator {
-  static RuntimeAllocator& current() noexcept;
-
-  RuntimeAllocator() = default;
-  RuntimeAllocator(size_t script_mem_size, size_t oom_handling_mem_size);
-
-  void init(void * buffer, size_t script_mem_size, size_t oom_handling_mem_size);
-  void free();
-
-  void * alloc_script_memory(size_t size) noexcept;
-  void * alloc0_script_memory(size_t size) noexcept;
-  void * realloc_script_memory(void *mem, size_t new_size, size_t old_size) noexcept;
-  void free_script_memory(void *mem, size_t size) noexcept;
-
-  void * alloc_global_memory(size_t size) noexcept;
-  void * alloc0_global_memory(size_t size) noexcept;
-  void * realloc_global_memory(void *mem, size_t new_size, size_t old_size) noexcept;
-  void free_global_memory(void *mem, size_t size) noexcept;
-
-  memory_resource::unsynchronized_pool_resource memory_resource;
-};
 
 struct KphpCoreContext {
   /**
